CASSANDRASC-36: Support for ErrorHandler in Sidecar
This commit adds a default `ErrorHandler` to the `Route#failureHandler()`. The default implementation
for Sidecar is the `JsonErrorHandler` implementation which is also provided as part of this commit.

The `JsonErrorHandler` class serializes `HttpException`s and `REQUEST_TIMEOUT`s to a JSON representation
that provides additional information to the response.

Additionally, we allow for custom implementations of the `ErrorHandler` to be injected (via module
overrides). This allows downstream projects to provide custom implementations of the `ErrorHandler` to
fit the specific needs of the project.