CASSANDRASC-46: Migrate minikube to testcontainers for integration tests
The existing Cassandra Sidecar integration testing suite uses Minikube to provision a
Cassandra service and it performs tests against the running database. This database runs
on Minikube.

A new alternative is to use [testcontainers](https://www.testcontainers.org), which requires
Docker to run tests. The concept is similar, but the benefit is that *testcontainers* is
well integrated into the java ecosystem and it works well with junit5. By replacing Minikube
with `testcontainers` we can simplify the setup process and reduce the complexity for running
integration tests.

Additionally, `testcontainers` is supported as part of the testing infrastructure inside
[CircleCI](https://www.testcontainers.org/supported_docker_environment/continuous_integration/circle_ci/).
Currently, our Minikube tests are broken both locally and in CircleCI. Moving to `testcontainers`
would also unblock us for further development.