Do not shade jackson-annotations
When shading jackson-annotations, extension libraries won't be able to annotate
POJOs with the unshaded jackson annotations (for example @JsonProperty), and
they would be forced to use the shaded annotation. In this commit, we exclude
jackson-annotations so that extension clients can keep annotating POJOs without
the need to use the shaded annotations.