CASSANDRASC-50: Deprecate the sidecar cassandra health endpoint containing instance segment
This commit deprecates the Cassandra Health endpoint containing the instance segment
in the path. This endpoint is currently unused and it is replaced by the health
endpoint with the `instanceId` query string parameter. Since the `instanceId` is optional
we move the path param (mandatory) to the query param (optional).

This commit also moves the CassandraHealthService from jax RS to a vertx Handler.
It also moves the HealthService to a inlined handler, simplifying the service.

patch by Francisco Guerrero; reviewed by Yifan Cai, Dinesh Joshi for CASSANDRASC-50