CASSANDRASC-53 Ignore unknown properties during Sidecar client deserialization
This commit modifies the way the `DecodableRequest` handles unknown properties in the
JSON payload. To support the evolution of the server API, we allow the Sidecar Client
to be more flexible when it encounters unknown properties, and we make it ignore these
new properties.

patch by Francisco Guerrero; reviewed by Dinesh Joshi, Yifan Cai for CASSANDRASC-53