CASSANDRASC-59: Expose JMX host and port from JMXClient
In this commit, we expose the JMX host and port from JMXClient to make it available
to implementations of the `ICassandraAdapter`.

This information can be valuable as different implementations and integrations need
to have this information available during the application execution.

patch by Francisco Guerrero; reviewed by Dinesh Joshi, Yifan Cai for CASSANDRASC-59