CASSANDRASC-64: File descriptor is not being closed on MD5 checksum
This commit fixes an issue where the file descriptors are not being closed during
the MD5 checksum during SSTable upload. This can potentially cause the JVM to run
out of available file descriptors during execution, and it would force an operator
to restart the Sidecar process to return to a working state.

In this commit, we ensure the file is closed after the checksum is calculated.

Additionally, this commit fixes a rare ConcurrentModificationException encountered
in `SSTableImporter` whree the `importQueuePerHost` does not use a thread-safe map.