CASSANDRASC-66 Fix builds in Apache CI
Currently, tests running under {{org.apache.cassandra.sidecar.HealthServiceSslTest}}
[are failing](https://ci-cassandra.apache.org/job/cassandra~sidecar/42/) when running
inside ASF's CI. Logs are showing that some resources (keystore and truststore) are
not found. This is causing the tests to fail.

In this commit, we read the resource from the stream, which is guaranteed to exist
as long as the resource exists and the resource name is correct, then we write the
resource to a temporary directory, and use the file name to set the keystore and
truststore path as part of the configuration options.

patch by Francisco Guerrero; reviewed by Dinesh Joshi, Yifan Cai for CASSANDRASC-66