CASSANDRASC-67 Fix relocation of native libraries for vertx-client-shaded
OpenSSL is currently unavailable in the vertx-client-shaded library. This is due to the
incorrect relocation of the netty native libraries under META-INF/native/libnetty.

In this commit we fix the relocation of the native libraries and we ensure that the
relocation is properly configured by adding tests that guarantee OpenSSL loads correctly
with the correct relocation.