CASSANDRASC-72 [CircleCI] Split unit tests and integration tests
As the number of integration tests continues to grow, it is desirable to split
unit tests and integration tests. It is also desirable to parallelize integration
tests, each integration test should be its own job. The goals of this improvement
are:

- Fail fast on checkstyle or minor errors
- Speed up test runtime by running integration tests in parallel
- Isolate failing tests to specific combinations of Cassandra

In this commit, we run unit tests individually for both java 8 and java 11. We split
integration tests into its own jobs, split per java version and cassandra version
combination.