Shade Jackson completely to prevent incompatibility issues
Before, we we shading almost all of Jackson but left the annotations unshaded/relocated. However, this causes problems
in older Spark environments where the annotations in the class path don’t quite match what the other shaded parts of
Jackson expect (missing classes being the most serious issue). This can cause the library to fail on certain Spark versions.
This commit removes the exclusion of the annotations project from shading in the vertx-client-shaded project.

- Additionally, tweaked CircleCI settings so that we don't run out of virtual memory and see the OS OOM kill the test exexutor.