CASSANDRASC-81 Improve TokenRangeReplicasResponse payload
The `TokenRangeReplicasResponse` returns a list with `ReplicaMetadata`. This information is used by clients to lookup
replica metadata. The lookup is done with the replica information, which consists of the `ip:port`.

Clients are looping over the `ReplicaMetadata` list and matching IP and port to retrieve the metadata object. Instead,
we can improve the payload by changing the data structure from a list to a map, and have clients lookup by the replica
(ip + port), without having to loop.

[CASSANDRASC-81](https://issues.apache.org/jira/browse/CASSANDRASC-81)