CASSANDRASC-83: Require gossip to be enabled for ring and token ranges mapping endpoints
The ring and token ranges mapping endpoints leverage JMX to produce the payload. Ring and token ranges
mapping information is derived from gossip information available to the Cassandra instance. If gossip
has been disabled, the information might provide an inconsistent view of the cluster which is
undesirable. We should instead return a 503 (Service Unavailable) to the client whenever the client
tries to access these endpoints and gossip has been disabled.

In this commit, we ensure the endpoints return a 503 (Service Unavailable) when gossip is disabled.

Patch by Francisco Guerrero; Reviewed by Dinesh Joshi, Josh McKenzie, Yifan Cai for CASSANDRASC-83