CASSANDRASC-76: Sidecar does not handle keyspaces and table names with mixed case
Cassandra Sidecar does not properly handle API requests for endpoints that have keypaces
or table names that need to be quoted, for example when names have mixed case or when
the name is a reserved keyword in Cassandra.

In this commit, we perform special handling when the keyspace or table names are quoted
in the path params. We add tests to ensure that handling is correct.

Additionally, we fix the validation for keyspaces and table names without quotes and add
special validation for quoted names.