CASSANDRASC-84 Expose additional node settings
Sidecar exposes settings from the Cassandra node via the node settings API endpoint. The information exposed is
limited, and we need to start exposing additional information from the `system.local` table, for example
`datacenter` information, owned token ranges, and the local address and port for the native protocol. This
information can be consumed by Sidecar itself, as well as the Cassandra Analytics library.

In this commit, we expose additional settings for the node.