CASSANDRASC-87: Add JMX health checks during the periodic health checks
In this commit, we add health checks based on the JMX connectivity to the managed
Cassandra instances. Additionally, we construct the NodeSettings object based on
JMX. This allows the Sidecar process to be able to determine an adapter for the
node even if the node is in joining state, or its binary port has been disabled.

Co-authored-by: Doug Rohrer <doug@therohrers.org>
Co-authored-by: Francisco Guerrero <frankgh@apache.org>

Patch by Doug Rohrer, Francisco Guerrero; Reviewed by Yifan Cai for CASSANDRASC-87