CASSANDRASC-97: Add support for additional checksum validation during SSTable upload
In this commit we add the ability to support additional checksum algorithms for verification
during SSTable uploads. We introduce the `ChecksumVerifierFactory` that allows configuring
different implementations of the `ChecksumVerifier` interface.

This commit also adds support for XXHash32 checksums. Clients can now send the XXHash32 checksum
instead of MD5. This would allow both the clients and server the flexibility to utilize a more
performant algorithm.

Patch by Francisco Guerrero; Reviewed by TBD for CASSANDRASC-97