CASSANDRASC-147: Expose vert.x filesystem options configuration
By default, vert.x will attempt to resolve files from the application classpath
when it is unable to find them in the local filesystem. Additionally, by default
vert.x will cache any files that it reads from the classpath into the local
filesystem.

For Sidecar, this optimization is unnecessary as Sidecar doesn't package anything
in the classpath that might be used while running the application.

In this commit, we disable this optimization by default, but expose configuration
options to tune these options on need-basis.

Patch by Francisco Guerrero; Reviewed by Yifan Cai for CASSANDRASC-147